[feat/staff-management] refactor input component
